Remove reliance on ActiveSupport::Concern.
We're really not doing anything complex, and removing an unnecessary dependency is a Good Thing.
