+++ lib/gutentag/active_record/class_methods.rb
@@ -1,0 +1,5 @@
+module Gutentag::ActiveRecord::ClassMethods
+  def tagged_with(*tags)
+    Gutentag::TaggedWithQuery.call self, tags
+  end
+end
