+++ lib/gutentag/engine.rb
@@ -2,6 +2,6 @@ class Gutentag::Engine < Rails::Engine
   engine_name :gutentag
 
   ActiveSupport.on_load :active_record do
-    include Gutentag::HasManyTags
+    extend Gutentag::HasManyTags
   end
 end
